internal: remove DiagnosticWithFix infra
